requirements.txt
blobconverter==1.4.3
certifi==2024.6.2
charset-normalizer==3.3.2
depthai==2.26.0.0
idna==3.7
numpy==1.26.4
opencv-contrib-python==4.10.0.82
pillow==10.3.0
PyYAML==6.0.1
requests==2.32.3
urllib3==2.2.1
watchdog==4.0.1